packages/data_handler/pyproject.toml
[project]
name = "data-handler"
version = "0.1.0"
description = "Data handling for LLM training - streaming datasets, packing, influence-based mixing"
authors = [{ name = "WrinkleFree Team" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "datasets>=2.18.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "einops>=0.7.0",
    "safetensors>=0.4.0",
    "accelerate>=0.27.0",
    "sentencepiece>=0.2.0",
    "tiktoken>=0.7.0",
    "datasketch>=1.6.0",  # For MinHash deduplication in influence functions
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
flash-attn = [
    "flash-attn>=2.5.0",
]
fp8 = [
    "transformer-engine>=1.0.0",
]
apollo = [
    "apollo-torch>=0.1.0",  # Memory-efficient optimizer (1/8 of AdamW memory)
]
muon = [
    "muon>=0.1.0",  # Muon optimizer - 2x compute efficiency vs AdamW
]
muonclip = [
    # muon-clip not yet on PyPI - install from source when available
    # MuonClip = Muon + QK-clipping for training stability (default optimizer)
]

[project.scripts]
data-handler-train = "data_handler.scripts.train:main"
data-handler-pretrain = "data_handler.scripts.pretrain:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_handler"]

# Include config YAML files in the package
[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
  "src/data_handler/**/*.py",
  "src/data_handler/configs/**/*.yaml",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "smoke: Quick smoke tests",
    "integration: Integration tests requiring GPU or external services",
    "slow: Slow tests that take >30s",
]

[tool.coverage.run]
source = ["src/data_handler"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
fail_under = 0

[tool.coverage.html]
directory = "htmlcov"
